backend/requirements.txt
fastapi==0.128.7
uvicorn==0.40.0
motor==3.7.1
pymongo==4.16.0
python-dotenv==1.0.1
pydantic[email]==2.12.5
